Fixed problem about not switching screens after google log in
it was failing because the app was trying to switch the screens while it was in the Google's AuthViewController, instead of our main controller. Since Google's view controller isn't part of our project hierarchy, it wasn't able to switch views.
The fix was to manually set the view to the main view controller before perfomring a segue